+++ packages/platform-cloudflare/test/worker.ts
@@ -12,6 +12,7 @@ import {
   DEPLOYMENT_ID,
   PRODUCER_PREFIX,
   fixtureReconcilerLayer,
+  maintenanceRaceFailpoint,
   makeTestBindings,
   runtimeEvictionFailpoint,
   storageEvictionFailpoint,
@@ -44,6 +45,7 @@ const baseOptions: ConversationObjectOptions = {
   toolReconciler: fixtureReconcilerLayer,
   storageFailpoint: storageEvictionFailpoint,
   runtimeFailpoint: runtimeEvictionFailpoint,
+  maintenanceFailpoint: maintenanceRaceFailpoint,
 };
 
 const possessionAuthorizationLayer = Layer.merge(