+++ packages/effect-agent/CHANGELOG.md
@@ -1,5 +1,14 @@
 # effect-agent
 
+## 0.1.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @effect-agent/capabilities@0.1.0-beta.15
+  - @effect-agent/core@0.1.0-beta.15
+  - @effect-agent/engine@0.1.0-beta.15
+
 ## 0.1.0-beta.14
 
 ### Patch Changes