+++ packages/pr-review/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @effect-agent/pr-review
 
+## 0.1.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - effect-agent@0.1.0-beta.15
+
 ## 0.1.0-beta.14
 
 ### Minor Changes